requirements.txt
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
html5lib==0.9999999
idna==2.6
Keras==2.0.8
Markdown==2.6.9
matplotlib==2.1.0
networkx==2.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-resize-image==1.1.11
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.4
scikit-image==0.13.1
scipy==0.19.1
six==1.11.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
urllib3==1.22
Werkzeug==0.12.2
